+++ rest_email_auth/serializers.py
@@ -199,6 +199,82 @@ class EmailVerificationSerializer(serializers.Serializer):
         return key
 
 
+class PasswordResetRequestSerializer(serializers.Serializer):
+    """
+    Serializer for requesting a password reset.
+    """
+    email = serializers.EmailField(
+        help_text=_("The email address to send the password reset to."))
+
+    def save(self):
+        """
+        Send out a password reset if the provided data is valid.
+
+        If the provided email address exists and is verified, a reset
+        email is sent to the address.
+
+        Returns:
+            The password reset token if it was returned and ``None``
+            otherwise.
+        """
+        try:
+            email = models.EmailAddress.objects.get(
+                email=self.validated_data['email'],
+                is_verified=True)
+        except models.EmailAddress.DoesNotExist:
+            return None
+
+        token = models.PasswordResetToken.objects.create(email=email)
+        token.send()
+
+        return token
+
+
+class PasswordResetSerializer(serializers.Serializer):
+    """
+    Serializer for reseting a user's password.
+    """
+    key = serializers.UUIDField(
+        help_text=_('The key received by the user in the password reset '
+                    'email.'),
+        write_only=True)
+    password = serializers.CharField(
+        help_text=_("The user's new password."),
+        style={'input_type': 'password'},
+        write_only=True)
+
+    def save(self):
+        """
+        Reset the user's password if the provided information is valid.
+        """
+        token = models.PasswordResetToken.objects.get(
+            key=self.validated_data['key'])
+
+        token.email.user.set_password(self.validated_data['password'])
+        token.email.user.save()
+
+        logger.info("Reset password for %s", token.email.user)
+
+        token.delete()
+
+    def validate_key(self, key):
+        """
+        Validate the provided reset key.
+
+        Returns:
+            The validated key.
+
+        Raises:
+            serializers.ValidationError:
+                If the provided key does not exist.
+        """
+        if not models.PasswordResetToken.valid_tokens.filter(key=key).exists():
+            raise serializers.ValidationError(
+                _("The provided reset token does not exist, or is expired."))
+
+        return key
+
+
 class RegistrationSerializer(serializers.ModelSerializer):
     """
     Serializer for registering new users.